+++ tests/testthat/tests.make_standata.R
@@ -238,11 +238,6 @@ test_that("make_standata handles multivariate models", {
 
 test_that("make_standata returns correct data for autocor structures", {
   dat <- data.frame(y = 1:10, x = rep(0, 10), tim = 10:1, g = rep(3:4, 5))
-  sdata <- make_standata(y ~ x, data = dat, autocor = cor_arr(~tim|g))
-  expect_equal(sdata$Yarr, cbind(c(0,9,7,5,3,0,10,8,6,4)))
-  
-  sdata <- make_standata(y ~ x, data = dat, autocor = cor_arr(~tim|g, r = 2))
-  expect_equal(sdata$Yarr, cbind(c(0,9,7,5,3,0,10,8,6,4), c(0,0,9,7,5,0,0,10,8,6)))
   
   sdata <- make_standata(y ~ x, data = dat, autocor = cor_ma(~tim|g))
   expect_equal(sdata$J_lag, as.array(c(1, 1, 1, 1, 0, 1, 1, 1, 1, 0)))